+++ showcase/integrations/langgraph-python/src/app/demos/reasoning-custom/reasoning-block.tsx
